📦 Move metadata-extract feature into subfolder and rename from metadatapreview
Relocate the 6 preview-only files into metadata-extract/, rename the
MetadataPreviewSection component to MetadataExtractSection, and rename the
RBAC scope app:works:metadatapreview -> app:works:metadata-extract (scope key
metadataPreview -> metadataExtract, hasMetadataExtractScope). No DB migration.
