+++ platform/scms/app/routes/app/works.$workId.upload.$workVersionId/metadata-extract/anthropic.server.ts
