+++ platform/scms/app/routes/app/works.$workId.upload.$workVersionId/metadata-extract/fetchPreviews.server.ts
